requirements.txt
jupyterlab
ipykernel
openpyxl
python-dotenv
numpy>=2.3.1
pandas
matplotlib
seaborn
plotly
statsmodels
scikit-learn

# Google client library
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib